refactor MailBuilderUtil
