+++ restservice/src/test/java/org/itevents/util/mail/MailBuilderUtilTest.java
@@ -1,0 +1,31 @@
+package org.itevents.util.mail;
+
+import org.itevents.model.Event;
+import org.itevents.test_utils.BuilderUtil;
+import org.junit.Test;
+import org.junit.runner.RunWith;
+import org.springframework.test.context.ContextConfiguration;
+import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
+
+import javax.inject.Inject;
+import javax.xml.bind.JAXBException;
+import javax.xml.transform.TransformerException;
+import java.io.IOException;
+import java.text.ParseException;
+import java.util.List;
+
+import static org.junit.Assert.assertEquals;
+
+@RunWith(SpringJUnit4ClassRunner.class)
+@ContextConfiguration(locations = {"classpath:applicationContext.xml", "classpath:applicationContextTestAddon.xml"})
+public class MailBuilderUtilTest {
+    @Inject
+    String expectedDigestEmail;
+
+    @Test
+    public void testMailBuild() throws JAXBException, ParseException, IOException, TransformerException {
+        List<Event> filteredEvents = BuilderUtil.buildEventsForMailUtilTest();
+        String returnedDigestEmail = new MailBuilderUtil().buildHtmlFromEventsList(filteredEvents);
+        assertEquals(expectedDigestEmail, returnedDigestEmail);
+    }
+}
