+++ restservice/src/test/resources/util/mail/weeklyDigestEmail.html
@@ -1,0 +1,64 @@
+<div xmlns:xalan="http://xml.apache.org/xslt">
+    <table cellpadding="0" cellspacing="0" width="100%" border="0" bgcolor="#f2f2f2">
+        <tr>
+            <td valign="top" align="center">
+                <table cellspacing="0" cellpadding="0" width="560" border="0" style="border-width:0px;border-color:#cccccc;border-style:solid;" bgcolor="#ffffff">
+                    <tr>
+                        <td align="center" width="560">
+                            <table border="0" cellspacing="0" width="560" cellpadding="0">
+                                <tr>
+                                    <td align="center" width="560">
+                                        <img height="100" src="" width="530" border="0" alt="Our logo"/>
+                                    </td>
+                                </tr>
+                            </table>
+                            <table width="560" border="0" cellspacing="0" cellpadding="0">
+                                <tr>
+                                    <td width="560" height="20"/>
+                                </tr>
+                            </table>
+                            <table width="520" border="0" cellspacing="0" cellpadding="0">
+                                <tr>
+                                    <td align="left" valign="top" style="color:#000000;font-family:Arial, Helvetica, sans-serif;line-height:17px;font-size:11px;">
+                                        <p style="margin:0 0 10px;">
+                                            <h3 style="margin:0;">Java</h3>
+                                            <span style="font-style: italic;">2115-07-10T00:00:00+03:00</span>
+                                            <br/>
+                                            <span>Beresteyska - java@gmail.com</span>
+                                            <br/>
+                                            <a href="http://www.java.com.ua">Register to Java</a>
+                                        </p>
+                                    </td>
+                                </tr>
+                            </table>
+                            <table width="520" border="0" cellspacing="0" cellpadding="0">
+                                <tr>
+                                    <td width="520" height="10"/>
+                                </tr>
+                            </table>
+                            <table width="520" border="0" cellspacing="0" cellpadding="0">
+                                <tr>
+                                    <td align="left" valign="top" style="color:#000000;font-family:Arial, Helvetica, sans-serif;line-height:17px;font-size:11px;">
+                                        <p style="margin:0 0 10px;">
+                                            <h3 style="margin:0;">Ruby</h3>
+                                            <span style="font-style: italic;">2115-07-20T00:00:00+03:00</span>
+                                            <br/>
+                                            <span>Shulyavska - ruby@gmail.com</span>
+                                            <br/>
+                                            <a href="http://www.ruby.com.ua">Register to Ruby</a>
+                                        </p>
+                                    </td>
+                                </tr>
+                            </table>
+                            <table width="520" border="0" cellspacing="0" cellpadding="0">
+                                <tr>
+                                    <td width="520" height="10"/>
+                                </tr>
+                            </table>
+                        </td>
+                    </tr>
+                </table>
+            </td>
+        </tr>
+    </table>
+</div>
